Changed the controller that handles song submission to be more safe and reasonable.
Now it accounts for the user not sending data, or sending the wrong data by checking each request against Last.fm (or the cache) to verify the song credentials, otherwise the user would be able to just send a post request with the artist and name fields set to whatever arbitrary text they wanted.

Also the MusicTrait has been altered to provide the necessary API calls to last.fm to make this happen.

MusicModel has been changed to return some fields of the submitted song when songSubmit is called, this is just to make more sense as a response to the POST request for song submission.